requirements.txt
requests
python-dotenv
certifi
schedule
pandas
pandas-stubs
streamlit
django-scheduler
pyyaml
altair
pytest
psutil